refactor(binder): extract standalone BinderGameDetailPanel; triage comments with reactions
